Combine runpath list tests
Also converts to pytest from unittest
